Create global js file which is used in all pages
